+++ src/main.js
@@ -1,10 +1,10 @@
+import 'normalize.css'
 import Vue from 'vue'
 import App from 'COMPONENTS/App'
 import router from 'ROUTERS/'
 import store from 'STORE/store'
 import './filters/'
-import 'babel-polyfill'
-
+import './public/ElementComponents'
 Vue.config.productionTip = false
 
 /* eslint-disable no-new */