+++ src/pages/Login.vue
@@ -1,22 +1,23 @@
 <template>
 <div class="login">
   <h2>{{ loginMsg }}</h2>
-  <div class="login-area">
-    <div class="form-group">
-      <input type="text" class="form-control" placeholder="用户名" autofocus v-model="user.usr">
-    </div>
-    <div class="form-group">
-      <input type="password" class="form-control" placeholder="密码" v-model="user.pwd">
-    </div>
-      <button class="btn btn-primary" @click="login" :disabled="loginValid">提交</button>
-      <button class="btn">取消</button>
-  </div>
+  <el-form :model="user">
+    <el-form-item>
+      <el-input v-model="user.usr" :placeholder="placeHoler.usr"></el-input>
+    </el-form-item>
+    <el-form-item>
+      <el-input v-model="user.pwd" :placeholder="placeHoler.pwd"></el-input>
+    </el-form-item>
+    <el-form-item>
+      <el-button plain type="primary" size="small" @click="login" :disabled="loginValid">提交</el-button>
+      <el-button plain size="small">取消</el-button>
+    </el-form-item>
+  </el-form>
 </div>
 </template>
 
 <script>
 import userService from 'SERVICES/userService'
-import * as types from 'STORE/types'
 
 export default {
   data () {
@@ -45,9 +46,9 @@ export default {
         const token = data.token
         const user = data.user
         if (token) {
-          this.$store.commit(types.LOGIN, token)
-          this.$store.commit(types.USER, user)
-          this.$router.push({ path: this.$route.query.redirect })
+          this.$store.commit('login', token)
+          this.$store.commit('user', user)
+          this.$router.push({ path: this.$route.query.redirect || '/home' })
         }
       })
     }